requirements.txt
allabolag
beautifulsoup4
builtwith
dnspython
googlesearch-python
pytesseract
#tesseract-ocr
python-whois
requests[socks]
PySocks
stem
GoogleNews
